refactor(chat): split stream registry into focused modules
Break the 900+ line stream-registry-service into types, state, broadcast,
persist, replay, and engine modules with a thin Effect service facade.
